Add files via upload
I've completed this project on a scholarship for independent journalists and activists from Eastern Europe at Forset, a great all-things-data NGO based in Tbilisi, Georgia. This didn't include any coding, but is still important to me as a project that taught me how to ask the right questions of data, even if using basic tools like Excel, and create great custom visualisations using "design" software (in this case – Adobe Illustrator). The project itself explores the problem of domestic violence against women in Russia and is based on data from independent sociological surveys as well as the UN and the World Bank data.