Add files via upload
I've completed this project on a scholarship for independent journalists and activists from Eastern Europe at Forset, a great all-things-data NGO in Tbilisi, Georgia. This pdf poster is dear to my heart since it was a great way to learn the fundamental skill of asking the rights questions of data, as well to learn how to combine various sources of data to tell a story, as well as how to create compelling custom data visualisations using 'designer' software (in this case Adobe Illustrator).